Move test for Guice injector out of the 'demo' package.
